# 51012 speed up module __init__.py thanks to scalene python profiler