--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,5 +1,9 @@
 # [CHANGELOG](https://keepachangelog.com/en/1.0.0/)
 
+## [5.10.12](https://github.com/gdsfactory/gdsfactory/pull/463)
+
+- speed up module gf/__init__.py thanks to scalene python profiler
+
 ## [5.10.8](https://github.com/gdsfactory/gdsfactory/pull/462)
 
 - fix documentation (add `pip install jaxlib jax`) to `make plugins`