- speed up module gf/__init__.py thanks to scalene python profiler, from 9 seconds to 4.5 seconds
- tests went down from 30s to 25s in laptop